+++ astropy/io/fits/scripts/fitsheader.py
@@ -42,13 +42,11 @@ header that describes the compression.
 With Astropy installed, please run ``fitsheader --help`` to see the full usage
 documentation.
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
+from __future__ import absolute_import, division, print_function
 
 import sys
 
 from ... import fits
-from .... import table
 from .... import log
 
 
@@ -133,17 +131,14 @@ class HeaderFormatter(object):
         for idx, hdu in enumerate(hdukeys):
             try:
                 cards = self._get_cards(hdu, keywords, compressed)
-
-                if idx > 0:  # Separate HDUs by a blank line
-                    result.append('\n')
-                result.append('# HDU {hdu} in {filename}:\n'.format(
-                              filename=self.filename,
-                              hdu=hdu
-                              ))
-                result.append('{0}\n'.format('\n'.join([str(c)
-                                                        for c in cards])))
             except ExtensionNotFoundException:
-                pass
+                continue
+
+            if idx > 0:  # Separate HDUs by a blank line
+                result.append('\n')
+            result.append('# HDU {} in {}:\n'.format(hdu, self.filename))
+            for c in cards:
+                result.append('{}\n'.format(c))
         return ''.join(result)
 
     def _get_cards(self, hdukey, keywords, compressed):
@@ -224,6 +219,7 @@ class TableHeaderFormatter(HeaderFormatter):
                 pass
 
         if tablerows:
+            from .... import table
             return table.Table(tablerows)
         return None
 
@@ -274,6 +270,7 @@ def print_headers_as_table(args):
     elif len(tables) == 1:
         resulting_table = tables[0]
     else:
+        from .... import table
         resulting_table = table.vstack(tables)
     # Print the string representation of the concatenated table
     resulting_table.write(sys.stdout, format=args.table)